refactor: extract playbook + hunt-suggestion/outcome routes into routes/playbookHunts.ts
Pure structural move out of createApp; no handler logic changed. Moves GET
/cases/:id/hunt-outcomes and the full /cases/:id/playbook* surface (control
get/put, list, sync, suggest-hunts, order, add, update, delete). Graduates
syncPlaybook + loadPlaybookControl onto RouteContext (hoisted functions) because
the staying POST /cases/:id/push/iris route also calls syncPlaybook; loadFreshHunts
and the notification side channel (dispatchNotify/caseLink) move into the module.
Reuses the already-graduated refreshVeloClients. Route inventory unchanged (316).
